remove unused cloning code and deps
