+++ packages/core/package.json
@@ -21,11 +21,8 @@
     "dotenv": "^6.2.0",
     "express": "^4.16.4",
     "fast-glob": "^2.2.6",
-    "glob-parent": "^3.1.0",
     "ignore": "^5.0.5",
     "mkdirp": "^0.5.1",
-    "node-fetch": "^2.3.0",
-    "npmi": "^4.0.0",
     "ora": "^3.0.0",
     "p-queue": "^6.0.2",
     "yarn": ">=1.17.3",
@@ -33,8 +30,7 @@
     "zero-builders-map": "^1.0.20-alpha.9",
     "zero-common-deps": "^1.0.20-alpha.9",
     "zero-dep-tree-js": "^1.0.14-alpha.0",
-    "zero-handlers-map": "^1.0.20-alpha.10",
-    "zero-static": "^1.0.10-alpha.8"
+    "zero-handlers-map": "^1.0.20-alpha.10"
   },
   "bundledDependencies": [],
   "repository": {
